ui setting page

+++ front-end/src/pages/OrderPage/Modals/ShowDeleteSelect.js
@@ -1,4 +1,12 @@
-function ShowDeleteSelect({ deleteId, setDeleteId, message, content, setShowDeleteSelect, handleDelete }) {
+function ShowDeleteSelect({
+  deleteId,
+  buttonContent,
+  setDeleteId,
+  message,
+  content,
+  setShowDeleteSelect,
+  handleDelete,
+}) {
   return (
     <div className="modal fixed top-0 animate-slideTopDown right-0 left-0 bottom-0 bg-black/30 z-[51] ">
       <div className="absolute py-6 px-6 top-[50%] translate-y-[-50%] overflow-hidden w-[420px] right-[50%] rounded-xl translate-x-[50%] bg-white">
@@ -7,8 +15,10 @@ function ShowDeleteSelect({ deleteId, setDeleteId, message, content, setShowDele
         <div className="flex items-center gap-6 mt-5">
           <div
             onClick={() => {
+              if (setDeleteId) {
+                setDeleteId('');
+              }
               setShowDeleteSelect(false);
-              setDeleteId('');
             }}
             className="bg-gray-300 px-6 cursor-pointer hover:opacity-80 transition-all py-2 rounded-lg"
           >
@@ -16,12 +26,14 @@ function ShowDeleteSelect({ deleteId, setDeleteId, message, content, setShowDele
           </div>
           <div
             onClick={() => {
+              if (handleDelete) {
+                handleDelete(deleteId);
+              }
               setShowDeleteSelect(false);
-              handleDelete(deleteId);
             }}
             className=" px-6 cursor-pointer py-2 rounded-lg hover:opacity-80 transition-all bg-orange-400 text-white"
           >
-            Delete
+            {buttonContent}
           </div>
         </div>
       </div>
